+++ src/generic/vacc.c
@@ -60,7 +60,7 @@ VEMBED(rcsid="$Id$")
 //
 // Author:   Nathan Baker 
 /////////////////////////////////////////////////////////////////////////// */
-VPUBLIC int Vacc_memChk(Vacc *thee) {
+VPUBLIC unsigned long int Vacc_memChk(Vacc *thee) {
     if (thee == VNULL) return 0;
     return Vmem_bytes(thee->vmem);
 }
